// frontend-portal-boa-entrega/src/app/pages/gsl/components/modulo-cadastral/depositos/models/enderecoDeposito.ts
export class EnderecoDeposito {
    id: string;
    logradouro: string;
    numero: string;
    complemento: string;
    bairro: string;
    cep: string;
    cidade: string;
    estado: string;
    depositoId: string;
}

export class CepConsulta {
    cep: string;
    logradouro: string;
    complemento: string;
    bairro: string;
    localidade: string;
    uf: string;
}